chore: update rules from meetup delete

+++ src/app/controllers/MeetupController.ts
@@ -36,13 +36,25 @@ class MeetupController {
 
   async index(req: Request, res: Response) {
     const meetups = await Meetup.findAll({ where: { user_id: req.userId } });
-    res.json(meetups);
+    return res.json(meetups);
   }
 
   async delete(req: Request, res: Response) {
     const { meetupId } = req.params;
-    const meetup = await Meetup.destroy({ where: { id: meetupId } });
-    res.json(meetup);
+    const meetup = await Meetup.findByPk(meetupId);
+
+    if (!meetup) {
+      return res.status(400).json({ error: 'meetup not found' });
+    }
+    if (meetup) {
+      // só pode cancelar meetup até 6h antes da data
+      if (isBefore(subHours(meetup.date, 6), new Date())) {
+        return res.status(400).json({ error: 'can only cancel meetup 6 hours in advance' });
+      }
+    }
+
+    await Meetup.destroy({ where: { id: meetupId } });
+    return res.send();
   }
 }
